requirements/prod.txt
asgiref==3.11.0
Django==6.0.1
sqlparse==0.5.5
python-decouple==3.8
